Refactor preset buttons and introduce state and signal managers
This update removes the preset buttons view and integrates it into the custom command view. It also introduces a state manager and a signal manager that handle app-wide flags and signals. Furthermore, the macro, serial, and tcp views have been renamed to macro, serial, and tcp tabs. These changes reflect a more suited application structure for the implementation of new features on the preset buttons and better overall app state management.