Convert methods to static in multiple services
Several methods across the system have been modified to be static as they don't rely on the instance state. These include `get_macro_directory` and `update_label_color` in the MacroController, `update_label_color` in StatusView, and `_parse_macro_commands` in MacroModel. No modifications to the other parts of the code are needed for this change.