Add flag_state_manager to macro_controller and update macro status
MacroController class is updated to include flag_state_manager in its constructor and check the 'serial_connected' status. MacroController now also signals when a macro is ready to run, considering the serial connection status. The MainWindow's handlers are updated to handle the macro running status and to update the macro status label accordingly. Added a method in flag_state_manager to get flag status. Lastly, the read-only attribute of macro_sequence_display is set to false in CommandView.