Implement serial communication functionality
Additional features were added to the system to allow it to send and receive data via serial communication. A new signal was emitted from the Command View when the single shot button was clicked. Upon this event, the Command Controller sends the constructed command to the Serial Controller. The Serial Controller then uses the Serial Model to send the command and to listen for data received from the serial port. Adjustments were also made to the Flag State Manager to reflect the current state of the system.