Add completed cycles update to GUI
A new signal to update the number of completed cycles in the GUI has been introduced. The updateCompletedCycles signal is emitted each time a cycle finishes, and is connected to the new method update_completed_cycles, which displays the current cycle count. Some adjustments to Macro View's layout have also been made to accommodate these changes.