Add ALARM_MESSAGE signal and refactor signal handling
A new signal named ALARM_MESSAGE has been added to the signal distributor. This led to significant refactoring of the alarm signal handling across various classes, which now utilise this new signal instead of the previous mechanism. This change has streamlined the handling of alarm messages and reduced redundancy in the code.