Remove preset button setup from CommandView
The setup and initialization of the preset buttons directly in the CommandView class has been removed. This simplifies the CommandView's initialization by taking out the arguments for button names. The preset button setup has been completely deleted for now, it would need to be implemented differently if it is required in the future.