Add signal to load command into view
A new PyQtsignal, LOAD_COMMAND_INTO_VIEW, was added to the SignalDistributor class to facilitate the loading of command details into the command view. The MacroController now emits this signal instead of calling the load_command_into_view function directly, with the MainWindow connecting to it to update the command view.