Add alarm logging and connect alarm signal
The alarm signal has been connected in the serial model which was previously commented out. In the serial reader, additional log messages are emitted whenever an alarm signals. Furthermore, the signature of the alarm signal in the signal distributor is modified to also accommodate a subcode alongside the alarm code.