Implement tab change listener in MainWindow
A signal-slot connection to the tab change event of the message_display_frame in MainWindow has been added. By adding a new method `on_tab_changed`, this update allows the flag_state_view to toggle between show and hide based on whether the Debug tab is selected or not. Additionally, a 'state updated' signal is emitted when entering debug mode.