Remove setup_parameters_line method
The setup_parameters_line method in command_service/command_view.py has been removed as it was no longer needed. The method was initiated to setup a line for input parameters which is now handled in a different way. A minor format adjustment was also made which does not change the logic of the programme.