Add new log file and update UI, .gitignore
A new log file has been created which contains detailed records of operations. UI changes have also been made in the main window to set the scrollbar of the log display to the maximum, enabling automatic scrolling to the end of the log messages. Also, an update is made to the .gitignore file to exclude all log files from version control.